Allow additional configuration for autocreated users
Right now feature of creating users aren't very useful, since
it doesn't allow to set additional parameters on the user and
most of developers have to create users in response callbacks.
With this simple fix, we return the power to the settings
CAS_AUTO_CREATE_USER, so users can set it and build their custom
middleware by inheritance.